Se crea el archivo app.ts para organizar el backend
